--- TOM/TOM.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29418.71
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TOM", "TOM.csproj", "{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM = Debug|ARM
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|ARM = Release|ARM
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM.ActiveCfg = Debug|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM.Build.0 = Debug|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM.Deploy.0 = Debug|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM64.Build.0 = Debug|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x64.ActiveCfg = Debug|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x64.Build.0 = Debug|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x64.Deploy.0 = Debug|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x86.ActiveCfg = Debug|x86
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x86.Build.0 = Debug|x86
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Debug|x86.Deploy.0 = Debug|x86
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM.ActiveCfg = Release|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM.Build.0 = Release|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM.Deploy.0 = Release|ARM
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM64.ActiveCfg = Release|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM64.Build.0 = Release|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|ARM64.Deploy.0 = Release|ARM64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x64.ActiveCfg = Release|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x64.Build.0 = Release|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x64.Deploy.0 = Release|x64
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x86.ActiveCfg = Release|x86
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x86.Build.0 = Release|x86
-		{6B6EDEBC-5F66-4AE2-83F6-A44E0D2D66C8}.Release|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {392AE723-13D6-44E6-915E-AC16499614D3}
-	EndGlobalSection
-EndGlobal
